// web/src/shared/api/generated/model/aiServiceConfigurationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SecretStatusResult } from './secretStatusResult';

export interface AiServiceConfigurationResult {
  llmProvider: string;
  chatModel: string;
  essayGradingModel: string;
  mathStepGradingModel: string;
  embeddingProvider: string;
  embeddingModel: string;
  transcriptionProvider: string;
  transcriptionModel: string;
  secrets: SecretStatusResult[];
}