// web/src/shared/api/generated/model/infrastructureConfigurationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AiServiceConfigurationResult } from './aiServiceConfigurationResult';
import type { AiServiceStatus } from './aiServiceStatus';
import type { IntegrationProviderResult } from './integrationProviderResult';
import type { SafetySwitchResult } from './safetySwitchResult';
import type { SecretStatusResult } from './secretStatusResult';

export interface InfrastructureConfigurationResult {
  environment: string;
  integrations: IntegrationProviderResult[];
  safetySwitches: SafetySwitchResult[];
  secrets: SecretStatusResult[];
  aiServiceStatus: AiServiceStatus;
  aiService: null | AiServiceConfigurationResult;
}