// web/src/shared/api/generated/model/runtimeSettingGroupResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { RuntimeSettingGroup } from './runtimeSettingGroup';
import type { RuntimeSettingResult } from './runtimeSettingResult';

export interface RuntimeSettingGroupResult {
  group: RuntimeSettingGroup;
  settings: RuntimeSettingResult[];
}